+++ src/sitePages/mainPage/ui/mainPage.tsx
